リトライ処理の見直しと環境センサークラスの分離
