+++ Pipfile
@@ -12,6 +12,8 @@ tqdm = "*"
 influxdb = "*"
 
 [dev-packages]
+pytest = "*"
+pytest-mock = "*"
 
 [requires]
 python_version = "3.7"
